add supabase auth

+++ src/app/api/tasks/[id]/route.js
@@ -1,31 +1,47 @@
 import { prisma } from "../../../../lib/prisma";
+import { createClient } from "../../../../../utils/supabase/server";
 
 export async function DELETE(req, { params }) {
   try {
+    const supabase = await createClient();
+    const {
+      data: { user },
+    } = await supabase.auth.getUser();
+
+    if (!user) {
+      return Response.json({ error: "Unauthorized" }, { status: 401 });
+    }
     const parameter = await params;
     const id = parseInt(parameter.id);
-
     const deletedTask = await prisma.task.delete({
-      where: { id },
+      where: { id, userId: user.id },
     });
     return Response.json(deletedTask);
   } catch (err) {
-    return Response.json({ message: "Failed to delete task" }, { status: 500 });
+    return Response.json({ error: err.message }, { status: 500 });
   }
 }
+
 export async function PATCH(req, { params }) {
   try {
+    const supabase = await createClient();
+    const {
+      data: { user },
+    } = await supabase.auth.getUser();
+
+    if (!user) {
+      return Response.json({ error: "Unauthorized" }, { status: 401 });
+    }
+
     const parameter = await params;
     const id = parseInt(parameter.id);
-    const request = await req.json();
+    const { completed } = await req.json();
     const updatedTask = await prisma.task.update({
-      where: { id },
-      data: {
-        completed: request.completed,
-      },
+      where: { id, userId: user.id },
+      data: { completed },
     });
     return Response.json(updatedTask);
   } catch (err) {
-    return Response.json({ message: "Failed to Patch task" }, { status: 500 });
+    return Response.json({ error: err.message }, { status: 500 });
   }
 }
